requirements.txt
appdirs==1.4.4
argcomplete==1.12.2
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.6
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
curlylint==0.12.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
-e git+https://github.com/jrief/django-admin-sortable2@86a44c759fc85f1a1ea87aaf78bd32b6daf7692b#egg=django_admin_sortable2
django-allauth==0.44.0
django-appconf==1.0.4
django-compressor==2.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-google-tag-manager==0.0.5
django-livereload-server==0.3.2
django-livesync==0.5
django-maintenance-mode==0.16.0
django-recaptcha==2.0.6
django-rename-app==0.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
escpos==1.8
Flask==1.1.2
future==0.18.2
geoip2==2.9.0
gitdb==4.0.7
GitPython==3.1.17
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libusb==1.0.23b7
MarkupSafe==1.1.1
maxminddb==2.0.3
oauthlib==3.1.0
parsy==1.1.0
pathspec==0.8.1
pep8==1.7.1
Pillow==8.1.2
pycodestyle==2.7.0
pycparser==2.20
PyESCPOS==0.3
pyflakes==2.2.0
PyJWT==2.0.1
pyserial==3.5
python-barcode==0.13.1
python3-openid==3.2.0
pytz==2021.1
pyusb==1.1.1
PyYAML==5.4.1
qrcode==6.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.15.0
smmap==4.0.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
Unidecode==1.2.0
urllib3==1.26.4
uuid==1.30
viivakoodi==0.8.0
watchdog==2.0.2
websocket-client==0.58.0
Werkzeug==1.0.1
wfastcgi==3.0.0
whitenoise==5.2.0
